Abstract an xml module so it can be used elsewere
